move tests into blas-tests
